test(models): resolve review coverage blockers
